Bogo_Sort.c
Bogo Sort is a highly inefficient and impractical sorting algorithm. It works by repeatedly shuffling the elements of an array until the array is sorted. The name "Bogo" is a play on the word "bogus," indicating that this sorting algorithm is not suitable for any practical use and is primarily used for educational purposes to illustrate the concept of a sorting algorithm